moved non methods into local.go changed GetCaller to global function
